Moved signals into their own package
